Fix issue for processing Patent files
Each record in a Patent file has a unique reference; as a result, we
may end up with using a hash array of the size of 300K, which is
extremely slow because it needs to dynamically allocate (keep doubling
buckets) memory for the hash.

Solution: since the references are unique, so just not use any hash
array for storing repeating references.